Reorganize nav, fix to top
Reorganizes the nav to a 3-part tabbed interface. There is also a home log on
the left (infinity sign) and a placeholder about link on the right.

Bulma doesn't support fixed position nav elements by default so this also
includes a hack to override default behavior for .nav.is-fixed elements.